+++ frontend/src/Settings/Indexers/IndexerSettingsConnector.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import { createSelector } from 'reselect';
-import { testAllIndexers } from 'Store/Actions/settingsActions';
+import { fetchIndexerOptions, fetchIndexers, testAllIndexers } from 'Store/Actions/settingsActions';
 import IndexerSettings from './IndexerSettings';
 
 function createMapStateToProps() {
@@ -15,6 +15,8 @@ function createMapStateToProps() {
 }
 
 const mapDispatchToProps = {
+  dispatchFetchIndexerOptions: fetchIndexerOptions,
+  dispatchFetchIndexers: fetchIndexers,
   dispatchTestAllIndexers: testAllIndexers
 };